commiting to merge

+++ search/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 class Offer(models.Model):
     seller_id = models.CharField(max_length=100)
-    price = models.DecimalField(max_digits=7, decimal_places=2) # there are no decimals in the price
+    price = models.IntegerField()
     isbn = models.CharField(max_length=20)
     course = models.CharField(max_length=100)
     condition = models.CharField(max_length=100)
